refactor: mcp own folder
