+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TodoListsModule } from './todo_lists/todo_lists.module';
 import { McpModule, McpTransportType } from '@rekog/mcp-nest';
 import { TodoItemsModule } from './todo_items/todo_items.module';
-import { ToolsModule } from './tools/tools.module';
+import { ToolsModule } from './mcp_tools_resources/tools/tools.module';
 
 @Module({
   imports: [
