+++ src/mcp_tools_resources/tools/get_items.tool.ts
