+++ src/mcp_tools_resources/tools/item_creation.tool.ts
